Split gatekeeper v1 route domains
